Aribitrarily typed buffers

+++ example/vect-add.cl
@@ -1,4 +1,4 @@
-__kernel void litmus_test(__global uint *a, __global uint *b, __global uint *c) {
+__kernel void litmus_test(__global uint *a, __global double *b, __global double *c) {
 	uint id = get_global_id(0);
-	c[id] = a[id] + b[id];
+	c[id] = (double) a[id] + b[id];
 }
